+++ NEWS.md
@@ -1,3 +1,11 @@
+### cholera 0.8.0.9214
+
+- use "_&_" as coordinate delimeter in pasteCoordsB().
+- add/set squareExitsB(latlong = FALSE).
+- amend squareCenterB() latlong variables.
+- amend nominal and latlong coordinates in Squares().
+
+
 ### cholera 0.8.0.9213
 
 - amend nominal and latlong coordinates in stJamesWorkhouse().